==> notebooks/00_setup.ipynb <==
# Setup Script for Necessary Directories

## Deps

import os

nltk_data_path = "../data/libs/nltk_data"
raw_data_path = "../data/raw"
processed_data_path = "../data/processed"
results_path = "../results"

os.makedirs(nltk_data_path, exist_ok=True)
os.makedirs(raw_data_path, exist_ok=True)
os.makedirs(processed_data_path, exist_ok=True)
os.makedirs(results_path, exist_ok=True)
