==> notebooks/01_load_and_merge_data.ipynb <==
# Combine/Merge All The Data from (https://zenodo.org/records/14227098)

## Merge per Month Data into One CSV

from src.utils.data_loader import load_or_merge_csvs

df = load_or_merge_csvs("../data/raw", "../data/raw/merged_tweets.csv")
df.head()

## Filter Out other Language and Save it

df_en = df[df['lang'] == 'en']

df_en.to_csv("../data/raw/merged_tweets_en.csv", index=False)

df_en.head()
